Ensure that the video doesn't disappear after locking the device
Fixes #36

Techinal details: MPMoviePlayerViewController observes UIApplicationDidEnterBackgroundNotification through -[MPMoviePlayerViewController _moviePlayerViewController_applicationDidEnterBackgroundNotification:]. This method calls _shouldContinuePlaybackInBackground. If it returns NO (which it will do unless the app has the audio `UIBackgroundModes` Info.plist key) then [moviePlayer stop] is called which causes the movie player view controller to be dismissed (if presented with `presentMoviePlayerViewControllerAnimated:`) or the video view to be removed from its superview (if presented with `presentInView:`).