Fix black screen after quick switch to background
When hitting the home button quickly after playing a video the `currentMoviePlayerController` is nil so it doesn’t get an associated `AVPlayer` object (PlayerKey) in `backgroundPlayback_applicationWillResignActive:`. When the application becomes active (`backgroundPlayback_applicationDidBecomeActive:`) the `currentMoviePlayerController` is not nil anymore but the player object is nil since it was not previously associated. We must ensure that we do not "restore" a nil player in order to avoid a black screen.