Wait for `start` to be called before canceling the operation
According to NSOperation documentation, this should not be required but we sometimes get this warning otherwise:
*** XCDYouTubeVideoOperation 0x7f8b18c84880 went isFinished=YES without being started by the queue it is in

When this warning occurs on Travis, it is always followed by this crash (diagnosed with zombies enabled):
*** -[NSKVONotifying_XCDYouTubeVideoOperation release]: message sent to deallocated instance 0x7fcde979c190

Ideally we should only wait for `start` to be called if the operation was added to a queue but there is no public API to know the queue an NSOperation is in. We could get it with [self valueForKeyPath:@"private._queue"] but this might break in the future so we always wait, even if the operation is not managed by an NSOperationQueue.